Created functions check_input, append_history and open_selection for duplicate code
check_input() : if in the selection an out of range number was entered, the program dies. If it is a valid ep. number / ep. range, it creates an episode list to watch/download.
append_history() : if there was an entry for the selected anime, it updates the entry, if not, it creates one for it.
open_selection() : opens all episodes in $episodes

These functions contain code that wasn't written by me, but was reused in the fnctionality I implemented. I created the functions to increase maintainability and decrease file size.